+++ utils/utils.h
@@ -34,7 +34,7 @@ char *string_trim(const char *);
 char *string_ltrim(const char *);
 char *string_rtrim(const char *);
 /* Functions working inplace */
-size_t string_format(char *, double, int, const char *);
+int string_format(char *, double, int, const char *);
 char *string_tokenize(char **, int);
 size_t string_length(const char *);
 size_t string_count(const char *, const char *);
